User registration, login, specific data and refresh tokens

+++ Tweetbook/Controllers/V1/PostsController.cs
@@ -1,14 +1,18 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Threading.Tasks;
 using Tweetbook.Contracts.V1;
 using Tweetbook.Contracts.V1.Requests;
 using Tweetbook.Contracts.V1.Responses;
 using Tweetbook.Domain;
+using Tweetbook.Extensions;
 using Tweetbook.Services;
 
 namespace Tweetbook.Controllers.V1
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class PostsController : Controller
     {
         private IPostService _postService;
@@ -38,12 +42,16 @@ namespace Tweetbook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest request)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+                return BadRequest(new { error = "You don't own this post" });
+
             var post = new Post
             {
                 Id = postId,
                 Name = request.Name
             };
-
             var updated = await _postService.UpdatePostAsync(post);
             return Ok(post);
         }
@@ -51,6 +59,11 @@ namespace Tweetbook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+                return BadRequest(new { error = "You don't own this post" });
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
@@ -62,15 +75,22 @@ namespace Tweetbook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post
+            {
+                UserId = HttpContext.GetUserId(),
+                Name = postRequest.Name
+            };
 
             await _postService.CreatePostAsync(post);
+            var response = new PostResponse { Id = post.Id };
+            return Created(UriLocation(post), response);
+        }
 
+        private string UriLocation(Post post)
+        {
             var baseUrl = HttpContext.Request.Scheme + "://" + HttpContext.Request.Host.ToUriComponent();
             var uriLocation = baseUrl + "/" + ApiRoutes.Posts.Get.Replace("{postId}", post.Id.ToString());
-
-            var response = new PostResponse { Id = post.Id };
-            return Created(uriLocation, response);
+            return uriLocation;
         }
     }
 }
